+++ include/ignition/rendering/Mesh.hh
@@ -19,6 +19,7 @@
 
 #include <map>
 #include <string>
+#include <unordered_map>
 #include <ignition/math/Matrix4.hh>
 #include "ignition/rendering/config.hh"
 #include "ignition/rendering/Geometry.hh"
@@ -55,9 +56,24 @@ namespace ignition
       /// * Map holding:
       ///     * Skeleton node names
       ///     * Local transformations of the skeleton nodes
-      /// \return True if all the transformations are set to the skeleton
       public: virtual void SetSkeletonLocalTransforms(
             const std::map<std::string, math::Matrix4d> &_tfs) = 0;
+
+      /// \brief Get skeleton node weight
+      /// \return Map of skeleton node name to its weight
+      /// * Map holding:
+      ///     * Skeleton node names
+      ///     * Weight the skeleton nodes
+      public: virtual std::unordered_map<std::string, float> SkeletonWeights()
+            const = 0;
+
+      /// \brief Set skeleton node weight
+      /// \param[in] _weights Map of skeleton node's name to its weight
+      /// * Map holding:
+      ///     * Skeleton node names
+      ///     * Weight the skeleton nodes
+      public: virtual void SetSkeletonWeights(
+            const std::unordered_map<std::string, float> &_weights) = 0;
 
       /// \brief Set whether a skeleton animation should be enabled or not
       /// \param[in] _name Name of animation