+++ ogre2/src/Ogre2MeshFactory.cc
@@ -186,6 +186,13 @@ bool Ogre2MeshFactory::LoadImpl(const MeshDescriptor &_desc)
       for (unsigned int i = 0; i < skel->AnimationCount(); ++i)
       {
         common::SkeletonAnimation *skelAnim = skel->Animation(i);
+        if (ogreSkeleton->hasAnimation(skelAnim->Name()))
+        {
+          ignwarn << "Skeleton animation not loaded. Animation with same "
+                  << "name already exists: " << skelAnim->Name() << std::endl;
+          continue;
+        }
+
         Ogre::v1::Animation *ogreAnim = ogreSkeleton->createAnimation(
             skelAnim->Name(), skelAnim->Length());